Get all activities for user
Currently printed for debugging. These will later be filtered for the right
bike and the durations summed.

I had to implement pagination. The default page size is 30 and can be a max
of 200; a page size of 100 seemed to perform well. It stops fetching when
the response is less than (but not equal to) the desired page size.

I had to reimplement `strava.NewStubResponseClient` because I needed a
client that was capable of responding to multiple requests.