Refactor tests to use gomega's ghttp
The mock transport was originally copied from the tests in the Strava
library. In hindsight, using Gomega's ghttp package is simpler. It makes the
responses and pagination for each test much more explicit.

I'm overriding the Dial functions to point the client at our ghttp server.
The tests are wrapped with some additional context blocks to give responses
to specific tests. The diff is easier to read without whitespace changes
(`-b`).